+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+* Changesets are only unique in the changelog filename + `id` + `author` combination, but so far `liquibase-opensearch`
+  considered them unique by just their `id`. New changesets are now stored with the correct combination as their ID.
+
 ## [0.2.0] - 2026-03-09
 
 ### Added